+++ src/table/column.rs
@@ -3,7 +3,7 @@ use crate::{types::*, value::*};
 /// Specification of a table column
 #[derive(Debug, Clone)]
 pub struct ColumnDef {
-    pub(crate) table: Option<DynIden>,
+    pub(crate) table: Option<TableRef>,
     pub(crate) name: DynIden,
     pub(crate) types: Option<ColumnType>,
     pub(crate) spec: Vec<ColumnSpec>,
@@ -36,6 +36,7 @@ pub enum ColumnType {
     JsonBinary,
     Uuid,
     Custom(DynIden),
+    Enum(String, Vec<String>),
 }
 
 /// All column specification keywords
@@ -399,6 +400,20 @@ impl ColumnDef {
         T: Iden,
     {
         self.types = Some(ColumnType::Custom(SeaRc::new(n)));
+        self
+    }
+
+    /// Set column type as enum.
+    pub fn enumeration<N, S, V>(&mut self, name: N, variants: V) -> &mut Self
+    where
+        N: ToString,
+        S: ToString,
+        V: IntoIterator<Item = S>,
+    {
+        self.types = Some(ColumnType::Enum(
+            name.to_string(),
+            variants.into_iter().map(|v| v.to_string()).collect(),
+        ));
         self
     }